+++ apps/backoffice-v2/src/domains/business-reports/components/AdsAndSocialMedia/AdsAndSocialMedia.tsx
@@ -1,5 +1,4 @@
 import React, { FunctionComponent } from 'react';
-import { RiskIndicators } from '@/common/components/molecules/RiskIndicators/RiskIndicators';
 import { Card } from '@/common/components/atoms/Card/Card';
 import { CardHeader } from '@/common/components/atoms/Card/Card.Header';
 import { CardContent } from '@/common/components/atoms/Card/Card.Content';
@@ -29,12 +28,14 @@ export const AdsAndSocialMedia: FunctionComponent<{
     link: string;
   }>;
   relatedAdsSummary: string;
-  relatedAdsImages: string[];
+  relatedAdsImages: Array<{
+    src: string;
+    link: string;
+  }>;
 }> = ({ violations, mediaPresence, adsImages, relatedAdsSummary, relatedAdsImages }) => {
   return (
     <div className={'space-y-8'}>
       <h3 className={'text-lg font-bold'}>Ads and Social Media Analysis</h3>
-      <RiskIndicators violations={violations} />
       <Card>
         <CardHeader className={'pt-4 font-bold'}>Social Media Presence</CardHeader>
         <CardContent className={'space-y-8'}>
@@ -93,8 +94,8 @@ export const AdsAndSocialMedia: FunctionComponent<{
         <CardContent className={'flex flex-col space-y-4'}>
           <div className={'grid grid-cols-[400px_400px] gap-8'}>
             {!!relatedAdsImages?.length &&
-              relatedAdsImages.map((src, index) => (
-                <AdExample key={src} src={src} alt={`Ad Example ${index + 1}`} />
+              relatedAdsImages.map(({ src, link }, index) => (
+                <AdExample key={src} src={src} link={link} alt={`Ad Example ${index + 1}`} />
               ))}
             {!relatedAdsImages?.length && <>No ads detected.</>}
           </div>