+++ apps/backoffice-v2/src/pages/Entity/pdfs/case-information/pages/RegistryInformationPage/types.ts
@@ -1,0 +1,6 @@
+export type IRegistryInformationItem = {
+  key: string;
+  title: string;
+  value: string;
+  valueType?: 'text' | 'link';
+};