+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,36 @@
 # kyb-app
 
+## 0.3.17
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+- Updated dependencies
+  - @ballerine/common@0.9.11
+  - @ballerine/blocks@0.2.5
+  - @ballerine/ui@0.5.5
+  - @ballerine/workflow-browser-sdk@0.6.17
+
+## 0.3.16
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/common@0.9.10
+  - @ballerine/blocks@0.2.4
+  - @ballerine/ui@0.5.4
+  - @ballerine/workflow-browser-sdk@0.6.16
+
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+  - @ballerine/workflow-browser-sdk@0.6.15
+
 ## 0.3.14
 
 ### Patch Changes