+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/workflow-core
 
+## 0.6.17
+
+### Patch Changes
+
+- Bump
+- Updated dependencies
+- Updated dependencies
+  - @ballerine/common@0.9.11
+
+## 0.6.16
+
+### Patch Changes
+
+- document changes
+- Updated dependencies
+  - @ballerine/common@0.9.10
+
+## 0.6.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.9
+
 ## 0.6.14
 
 ### Patch Changes